senior_project /IssacsIndex/requirements.txt
aniso8601==9.0.1
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cryptography==3.4.8
exceptiongroup==1.1.3
Flask==2.3.3
Flask-RESTful==0.3.10
flask-restx==1.1.0
Flask-SQLAlchemy==3.1.1
greenlet==2.0.2
gunicorn==20.1.0
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
iniconfig==2.0.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.3
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.26.2
oauthlib==3.2.0
packaging==23.2
playwright==1.38.0
pluggy==1.3.0
psycopg2-binary==2.9.9
pycairo==1.20.1
pyee==9.0.4
PyJWT==2.3.0
pyparsing==2.4.7
pytest==7.4.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
SecretStorage==3.3.1
serpapi==0.1.5
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.21
tomli==2.0.1
urllib3==2.0.6
wadllib==1.3.6
Werkzeug==2.3.7
zipp==1.0.0